fix: address follow-up self-upgrade review
