+++ tests/self_upgrade_helpers.py
@@ -4,18 +4,10 @@ These helpers patch subprocess, PATH lookup, and release-tag resolution so
 the focused test modules stay isolated from the real environment.
 """
 
-import errno
-import importlib.metadata
-import json
-import os
 import subprocess
-import urllib.error
-from unittest.mock import patch
 
 from typer.testing import CliRunner
 
-import specify_cli
-from specify_cli import app
 from specify_cli._version import (
     _InstallMethod,
     _UpgradePlan,
@@ -35,18 +27,9 @@ __all__ = (
     "_completed_process",
     "_detect_install_method",
     "_verify_upgrade",
-    "app",
-    "errno",
-    "importlib",
-    "json",
     "mock_urlopen_response",
-    "os",
-    "patch",
     "runner",
-    "specify_cli",
     "strip_ansi",
-    "subprocess",
-    "urllib",
 )
 
 runner = CliRunner()
